backend/requirements.txt
decimer==2.7.1
beautifulsoup4==4.12.3
faiss-cpu==1.8.0.post1
fastapi==0.115.0
MolScribe==1.1.1
numpy<2
opencv-python==4.9.0.80
Pillow==10.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
pymupdf==1.24.10
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
sqlmodel==0.0.16
torch==2.2.2
transformers==4.44.2
uvicorn[standard]==0.30.6
